refactor(daemon): split lifecycle/service.go into focused files
Split the ~1200-line service.go into 7 focused files:
- service.go: Service struct, New(), Options, core state management
- install.go: Install logic
- start_stop.go: Start/Stop logic
- upgrade.go: Upgrade logic, version helpers, backup
- memory.go: Memory attachment methods
- audit.go: Audit buffer and logging
- helpers.go: Internal helper functions

No logic changes — pure file reorganization. All tests pass.
